pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crq"
version = "0.1.0"
description = "Cosmic-ray / earthquake correlation toolkit — Homola et al. 2023 replication"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "polars>=0.19",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=13.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
